fix: Correct register reading and conversion from 2 byte to 1

+++ lis2ds12_reg.h
@@ -920,7 +920,7 @@ float_t lis2ds12_from_fs4g_to_mg(int16_t lsb);
 float_t lis2ds12_from_fs8g_to_mg(int16_t lsb);
 float_t lis2ds12_from_fs16g_to_mg(int16_t lsb);
 
-float_t lis2ds12_from_lsb_to_celsius(int16_t lsb);
+float_t lis2ds12_from_lsb_to_celsius(int8_t lsb);
 
 typedef struct
 {
@@ -989,7 +989,7 @@ int32_t lis2ds12_acceleration_module_raw_get(const stmdev_ctx_t *ctx,
                                              uint8_t *buff);
 
 int32_t lis2ds12_temperature_raw_get(const stmdev_ctx_t *ctx,
-                                     uint8_t *buff);
+                                     int8_t *buff);
 
 int32_t lis2ds12_acceleration_raw_get(const stmdev_ctx_t *ctx,
                                       int16_t *val);
